修复了新怪物有时不能正常刷小怪的问题 初步制作了第二种模式
git-svn-id: https://viruscollector/svn/Spacecraft@227 6ae6004f-9e4d-49be-878d-5da08febfbad

+++ src/spacecraftelements/Enemy/Slime.java
@@ -21,19 +21,23 @@ public class Slime extends Enemy {
 		super.volume = 40;
 		super.imageID = "Images//enemy//slime.png";
 		super.imagesize = 20;
-		super.getscore=9;
+		super.getscore = 9;
 	}
 
 	public boolean update() {
 		if (x == MainGame.bm.getShip().x && y == MainGame.bm.getShip().y) {
 			return false;
 		}
-		if(hit==true){this.imageID = "Images//enemy//slime1.png";hit=false;}
-		else {this.imageID = "Images//enemy//slime.png";}//hint when hit
+		if (hit == true) {
+			this.imageID = "Images//enemy//slime1.png";
+			hit = false;
+		} else {
+			this.imageID = "Images//enemy//slime.png";
+		}// hint when hit
 		int x1, y1;
 		double ratiox, ratioy, third;
-		x1 = MainGame.bm.getShip().x - x;
-		y1 = MainGame.bm.getShip().y - y;
+		x1 = (int) (MainGame.bm.getShip().x - x);
+		y1 = (int) (MainGame.bm.getShip().y - y);
 		third = Math.sqrt((double) x1 * x1 + y1 * y1);
 		ratiox = ((double) x1) / third;
 		ratioy = ((double) y1) / third;
@@ -46,25 +50,22 @@ public class Slime extends Enemy {
 
 	@Override
 	public SpecialEffect deathwhisper() {
-		double r=Math.random();
+		double r = Math.random();
 		if (r < 0.05) {
 			MainGame.bm.add(new W_StarWeapon(x, y));
-		} else if(r>0.15&&r<0.2){
+		} else if (r > 0.15 && r < 0.2) {
 			MainGame.bm.add(new H_bulletrain(x, y));
-		}
-		else if(r>0.2&&r<0.25){
+		} else if (r > 0.2 && r < 0.25) {
 			MainGame.bm.add(new S_speedup(x, y));
-		}
-		else if(r>0.3&&r<0.4){
+		} else if (r > 0.3 && r < 0.4) {
 			MainGame.bm.add(new S_repair(x, y));
-		}
-		else if(r>0.4&&r<0.5){
+		} else if (r > 0.4 && r < 0.5) {
 			MainGame.bm.add(new H_bulletblast(x, y));
-		}else if (r > 0.55 && r < 0.6) {
+		} else if (r > 0.55 && r < 0.6) {
 			MainGame.bm.add(new W_Shotgun(x, y));
-		}else if (r > 0.6 && r < 0.61) {
+		} else if (r > 0.6 && r < 0.61) {
 			MainGame.bm.add(new W_BasicWeapon(x, y));
 		}
-		return new SmallBlast(x,y);
+		return new SmallBlast(x, y);
 	}
 }
